Allow simpler service ENV setting for startup script
The `PLEK_` params are long, hard to remember, easy to typo.

Keep 'use-remote' versions of `static` / `content-store` as the default,
but allow them to be set to local versions using simpler/shorter ENV vars.

This also encapsulates the local paths for `static` / `content_store` to
avoid having to know the ports / type full addresses.