Add hard coded content item for prototype
So this can run on Heroku without needing to point at another
instance of content schemas, on a branch with the local
transaction example.

The controller code is pretty generic (if inelegant) and can
be re-used, as I suspect this is going to be a common pattern
for prototype branches. The location/naming of the hard coded
content item needs a bit more thought before merging this
into master